backend/app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pytesseract
pdfplumber
pdf2image
python-multipart

